### 2019-01-29 #1218 @whedon
[PRE REVIEW]: crystal-torture: A crystal tortuosity module
**Submitting author:** @connorourke (<a href="http://orcid.org/0000-0002-0703-8234">Conn O&#39;Rourke</a>)
**Repository:** <a href="https://github.com/connorourke/crystal_torture" target ="_blank">https://github.com/connorourke/crystal_torture</a>
**Version:** v1.0.3
