### 2019-01-29 #1218 @katyhuff
Sorry -- that darn close and comment button is the worst UI in the universe.
