### 2019-01-29 #1218 @katyhuff
@connorourke and @bjmorgan : Thank you for your submission!

I'm going to be looking for reviewers and could use your help. If you see anyone in [this list](https://bit.ly/joss-reviewers) who would be appropriate, I would be grateful if you could point them out. I'll look closely as well, but frequently the authors have insight on the appropriateness of reviewers that an editor cannot have. It's important that reviewers have domain knowledge in the field of the submission (computational physics/chemistry/materials)  as well as familiarity with the systems and languages involved (python, fortran, openMP).

Before I assign anyone,  I am going to do a little pre-review. You may see some issues pop up in your repo.
