### 2019-01-29 cbd85cb
Possible fix for compilation errors
