+++ setup.py
@@ -19,23 +19,23 @@ def setup_tort_ext(args,parent_package='',top_path=''):
     tort_src = [join('crystal_torture/','tort.f90')]
 
     config.add_library('_tort', sources=tort_src,
-                           extra_f90_compile_args = [ args["compile_args"]],
-                           extra_link_args=[args["link_args"]])
+                           extra_f90_compile_args = args["compile_args"],
+                           extra_link_args=args["link_args"])
 
     sources = [join('crystal_torture','f90wrap_tort.f90')]
     
     config.add_extension(name='_tort',
                           sources=sources,
-                          extra_f90_compile_args = [ args["compile_args"]],
-                          extra_link_args=[args["link_args"]],
+                          extra_f90_compile_args = args["compile_args"],
+                          extra_link_args=args["link_args"],
                           libraries=['_tort'],
                           include_dirs=['build/temp.' + return_include_dir()])
 
     dist_src = [join('crystal_torture/','dist.f90')]
     config.add_extension(name='dist',
                           sources=dist_src,
-                          extra_f90_compile_args = [ args["compile_args"]],
-                          extra_link_args=[args["link_args"]])
+                          extra_f90_compile_args = args["compile_args"],
+                          extra_link_args=args["link_args"])
 
 
     return config
@@ -58,24 +58,24 @@ def check_f2py_compiler():
      
     if 'GNU' in str(result):
         print('Found gnu compiler. Setting OpenMP flag to \'-fopenmp\'')
-        compile_args = '-fopenmp -lgomp -O3'
-        link_args = '-lgomp'
+        compile_args = ['-fopenmp', '-lgomp', '-O3']
+        link_args = ['-lgomp']
     elif 'Intel' in str(result):
         print('Found intel compiler. Setting OpenMP flag to \'-openmp\'')
-        compile_args = '-openmp -O3'
-        link_args = ''
+        compile_args = ['-openmp', '-O3']
+        link_args = []
     elif 'Portland' in str(result):
         print('Found portland compiler. Setting OpenMP flag to \'-mp\'')
-        compile_args = '-mp -O3'
-        link_args = ''
+        compile_args = ['-mp', '-O3']
+        link_args = []
     elif 'NAG' in str(result):
         print('Found NAG compiler. Setting OpenMP flag to \'-openmp\'')
-        compile_args = '-openmp -O3'
-        link_args = '' 
+        compile_args = ['-openmp', '-O3']
+        link_args = [] 
     else:   
         print('Not sure what compiler f2py uses. Speculatively setting OpenMP flag to \'-openmp\'')
-        compile_args = '-openmp -O3'
-        link_args = ''
+        compile_args = ['-openmp', '-O3']
+        link_args = [] 
 
     args = {'link_args':link_args,'compile_args':compile_args}
 
