+++ requirements.txt
@@ -1,5 +1,5 @@
-ddt
 coverage
+ddt
 numpy
-pymatgen
 f90wrap
+pymatgen
