### 2019-01-30 e704329
specify numpy version
