### 2019-01-30 1f16599
Check remove numpy version for docs build
