### 2019-01-31 7bf643f
Update version number in docs conf.py
