+++ docs/source/conf.py
@@ -50,7 +50,7 @@ author = 'Conn O\'Rourke'
 # The short X.Y version
 version = ''
 # The full version, including alpha/beta/rc tags
-release = '1.0.0'
+release = '1.0.4'
 
 
 # -- General configuration ---------------------------------------------------
