### 2019-01-31 314d630
Added check on python and replaced direct call to f2py in setup with numpy.distutils.fcompiler.get_default_fcompiler - avoids hanging due to bug in numpy 1.16.0
