### 2019-01-31 6785462
mock pymatgen for autodocs
