### 2019-03-04 #1218 @katyhuff
Dear reviewers,
This submission is in need of expertise in computational physics/chemistry/materials as well as familiarity with the systems and languages involved (python, fortran, openMP).

@goerz : Michael, are you able to and interested in taking on this review?
@jwuttke : Joachim, are you able to and interested in taking on this review?
@jochym : Pawel, are you able to and interested in taking on this review?
@vahtras: Olav, are you able to and interested in taking on this review?

- Title: crystal-torture: A crystal tortuosity module
- Summary: crystal_torture is a Python, Fortran and OpenMP crystal structure analysis module. The module contains a set of classes that enable: a crystal structure to be converted into a graph for network analysis, connected clusters of crystal sites (nodes) to be retrieved and output, periodicity of connected clusters of crystal sites to be determined, and relative path tortuosity to traverse a crystal within a connected cluster to be calculated for each site.
- Article Proof: https://github.com/openjournals/joss-papers/blob/joss.01218/joss.01218/10.21105.joss.01218.pdf
- Submitting author: @bjmorgan (Benjamin Morgan)
- Repository: https://github.com/connorourke/crystal_torture
- Version: v1.0.3

A very nice jupyter notebook example is found here: https://nbviewer.jupyter.org/github/connorourke/crystal_torture/blob/master/examples/crystal_torture_examples.ipynb
