### 2019-03-07 #1218 @katyhuff
Hi Folks,

@jwuttke : Joachim, are you able to and interested in taking on this review?
@jochym : Pawel, are you able to and interested in taking on this review?
@yxqd : Jiao, are you able to and interested in taking on this review?
@mkhorton : Matthew, are you able to and interested in taking on this review?

- Title: crystal-torture: A crystal tortuosity module
- Summary: crystal_torture is a Python, Fortran and OpenMP crystal structure analysis module. The module contains a set of classes that enable: a crystal structure to be converted into a graph for network analysis, connected clusters of crystal sites (nodes) to be retrieved and output, periodicity of connected clusters of crystal sites to be determined, and relative path tortuosity to traverse a crystal within a connected cluster to be calculated for each site.
- Article Proof: https://github.com/openjournals/joss-papers/blob/joss.01218/joss.01218/10.21105.joss.01218.pdf
- Submitting author: @bjmorgan (Benjamin Morgan)
- Repository: https://github.com/connorourke/crystal_torture
- Version: v1.0.3
