### 2019-03-08 #1306 @whedon
[REVIEW]: crystal-torture: A crystal tortuosity module
**Submitting author:** @connorourke (<a href="http://orcid.org/0000-0002-0703-8234">Conn O&#39;Rourke</a>)
**Repository:** <a href="https://github.com/connorourke/crystal_torture" target ="_blank">https://github.com/connorourke/crystal_torture</a>
**Version:** v.1.0.08

- [x] **Version:** v.1.0.08
