+++ README.md
@@ -1,4 +1,5 @@
 
+[![status](http://joss.theoj.org/papers/c3d8e702ecfee04f16a0ad6f14d96419/status.svg)](http://joss.theoj.org/papers/c3d8e702ecfee04f16a0ad6f14d96419)
 [![PyPI version](https://badge.fury.io/py/crystal-torture.svg)](https://badge.fury.io/py/crystal-torture)
 [![Build Status](https://travis-ci.com/connorourke/crystal_torture.svg?token=nTMqYYEUasQRTBsU6oCc&branch=master)](https://travis-ci.com/connorourke/crystal_torture)
 [![Coverage Status](https://coveralls.io/repos/github/connorourke/crystal_torture/badge.svg?branch=master)](https://coveralls.io/github/connorourke/crystal_torture?branch=master)
@@ -9,14 +10,12 @@
 
 
 ### **crystal_torture:** 
- `crystal_torture` is a Python, Fortran and OpenMP crystal structure analysis module. The module contains a set of classes
-that enable:
+ `crystal_torture` is a Python, Fortran and OpenMP crystal structure analysis module. The module contains a set of classes that enable:
 
-* a crystal structure to be converted into a graph for network analysis 
-* connected clusters of crystal sites (nodes) to be retrieved and output
-* periodicity of connected clusters of crystal sites to be determined
-* relative path tortuosity to traverse a crystal within a connected cluster to be calculated for each site
- 
+* a crystal structure to be converted into a graph for network analysis. 
+* connected clusters of crystal sites (nodes) to be retrieved and output.
+* periodicity of connected clusters of crystal sites to be determined.
+* relative path tortuosity to traverse a crystal within a connected cluster to be calculated for each site.
  
 ## Installation
 
