### 2019-03-17 ccc8a4a
Proposed fix for numpy requirements to run setup
