### 2019-03-17 ece367c
Fix for issue https://github.com/connorourke/crystal_torture/issues/4
