### 2019-03-17 01904fe
Replaced structure.species_and_occu with structure.species
