### 2019-03-28 #1306 @mkhorton
@connorourke, just a small note that your PyPI formatting doesn't seem to be working correctly, see [here](https://packaging.python.org/guides/making-a-pypi-friendly-readme/)
