### 2019-03-28 #1306 @mkhorton
@connorourke the only other checklist item missing is a guideline for contributions.

From a functionality point of view, I'm happy to recommend this software, and all functionality seems to be present and correct. I think this will be a very useful tool for the community (and, indeed, having studied percolation myself in the past, wish I had had this tool available!)

From a code point of view, I would have a few recommendations that, under my understanding of the JOSS guidelines, have no bearing on the review but I include here just as suggestions:

* Code generally looks to be of a good quality, but is not [PEP8](https://www.python.org/dev/peps/pep-0008) compliant. I would recommend passing it through a tool such as [Black](https://github.com/ambv/black) to re-format appropriately.

* The correct usage of this code requires explicitly running `graph.torture()` before other commands such as `graph.minimal_clusters` can be run. I would strongly recommend adding appropriate error messages so that if a user tries to run `graph.minimal_clusters` *before* `graph.torture()` it will give a hint that they need to do this (currently it gives `'NoneType' object is not iterable`).

* The paper/examples would be greatly improved if they generated a known-true example from literature. From my understanding, the output will be very dependent on the size of supercell used, and extrapolation of multiple supercell values will be needed to get correct values for an infinite cell? This has no bearing on this code itself which, as far as I can tell, correctly calculates clusters and tortuosity for the cells provided, but is relevant for the broader application of this code.

* There is a concept in pymatgen of transformations including, for example, an OrderDisorderTransformation to create ordered (not random) supercells. The more idiomatic (from a pymatgen point of view) way to construct the supercell for your spinel example would be to first create a *disordered* structure, i.e. to replace Mg with `{'Mg':0.9, 'Li': 0.05, 'Al': 0.05}`. To create a random supercell would then require creating a new `RandomOrderDisorderTransformation` class or similar (something pymatgen is definitely in need of). I only mention this to say that using site property labels to perform the doping is perfectly fine, but there might be a better/more idiomatic way of doing it that would also be useful for the broader community of people who use pymatgen.

* A cut-off radius approach for calculating the graph is not sufficient for all structures. There are always edge cases; this is especially true for, for example, the output of MD simulations. There's a whole collection of different `NearNeighbor` classes in pymatgen that determine whether two sites should be 'connected', and also a `StructureGraph` class to construct graphs using these strategies. However, in your case, I think the use of cut-off radius makes sense because it's *fast* (the pymatgen classes NearNeighbor classes are probably too slow for your use case). I only mention this point just because I've encountered issues myself before related to cut-off radii approaches that ultimately led to incorrect analysis outputs.

In summary however, I'm very excited about this code, and really look forward to see how it develops going forwards. I will add an additional comment to this thread to formally complete the review once all checklist items are completed.
