### 2019-04-04 #1306 @connorourke
@mkhorton,

Thanks for taking the time to do such a thorough review of the code, and for the useful suggestions. I'll make the changes you suggested, sort out the PyPi formatting issue and cross off the final checklist items asap.

I'd agree that the examples would benefit from a known-true example being included - and you are right that the results are dependent upon the size of supercell being used. We are currently preparing a publication which uses the code, and this is a point which we discuss within it - once it is published we will update the examples in the repo accordingly.

Thanks!
