### 2019-04-04 #1306 @mkhorton
Great! Looking forward to seeing it :)

Just to re-iterate, the additional comments don't have a bearing on my JOSS review (as I understand the JOSS guidelines). It's really the user's responsibility if they correctly extrapolate with supercell size etc. As long as the code itself works, is tested, documented, etc. and fulfills the other JOSS criteria, which it looks like it does, then I'm happy to recommend publication. The known-true example would be good however.
