### 2019-04-29 61dd0e3
Added README.rst for pypi description rendering
