+++ crystal_torture/version.py
@@ -1,1 +1,1 @@
-__version__ = '1.0.5'
+__version__ = '1.0.05'
