### 2019-04-30 9c94ac0
fix for pymatgen 2019.4.11, and >= included in install_requires
