+++ crystal_torture/pymatgen_interface.py
@@ -13,6 +13,7 @@ import time
 
 "Functions for setting up a node, cluster and graph using pymatgen"
 
+
 def map_index(uc_neighbours, uc_index, x_d, y_d, z_d):
     """
     Takes a list of neighbour indices for sites in the original unit cell, 
@@ -38,13 +39,19 @@ def map_index(uc_neighbours, uc_index, x_d, y_d, z_d):
     count = -1
     neigh = []
     append = neigh.append
-    for i,index in enumerate(uc_index):
-       for x in range(0,x_d,1):
-         for y in range(0,y_d,1):
-            for z in range(0,z_d,1):
-               count+=1
-               append([dist.shift_index(neighbour,[x,y,z]) for neighbour in uc_neighbours[i]])
+    for i, index in enumerate(uc_index):
+        for x in range(0, x_d, 1):
+            for y in range(0, y_d, 1):
+                for z in range(0, z_d, 1):
+                    count += 1
+                    append(
+                        [
+                            dist.shift_index(neighbour, [x, y, z])
+                            for neighbour in uc_neighbours[i]
+                        ]
+                    )
     return neigh
+
 
 def get_all_neighbors_and_image(structure, r, include_index=False):
 
@@ -101,12 +108,15 @@ def get_all_neighbors_and_image(structure, r, include_index=False):
         all_within_r = np.bitwise_and(all_dists <= r, all_dists > 1e-8)
 
         for (j, d, within_r) in zip(indices, all_dists, all_within_r):
-            nnsite = PeriodicSite(structure[j].species_and_occu, coords[j],
-                                    latt, properties=structure[j].properties,
-                                    coords_are_cartesian=True)
+            nnsite = PeriodicSite(
+                structure[j].species_and_occu,
+                coords[j],
+                latt,
+                properties=structure[j].properties,
+                coords_are_cartesian=True,
+            )
             for i in indices[within_r]:
-                item = (nnsite, d[i], j, image) if include_index else (
-                    nnsite, d[i])
+                item = (nnsite, d[i], j, image) if include_index else (nnsite, d[i])
                 neighbors[i].append(item)
     return neighbors
 
@@ -134,15 +144,17 @@ def create_halo(structure, neighbours):
 
     no_sites = len(structure.sites)
     for i in range(no_sites):
-       neighbours[i]=[dist.shift_index((27*neighbour[2]),neighbour[3]) for neighbour in neighbours[i]]
-      
+        neighbours[i] = [
+            dist.shift_index((27 * neighbour[2]), neighbour[3])
+            for neighbour in neighbours[i]
+        ]
+
     uc_index = [((site * 27)) for site in range(len(structure.sites))]
-    structure.make_supercell([x,y,z])
-    neighbours = map_index(neighbours,uc_index,x,y,z)
-    
-     
+    structure.make_supercell([x, y, z])
+    neighbours = map_index(neighbours, uc_index, x, y, z)
 
     return structure, neighbours
+
 
 def nodes_from_structure(structure, rcut, get_halo=False):
     """
@@ -158,41 +170,48 @@ def nodes_from_structure(structure, rcut, get_halo=False):
         - nodes (set(Nodes)): set of Node objects representing structure sites
   
     """
-    structure.add_site_property("UC_index", [str(i) for i in range(len(structure.sites))] )
-    neighbours = get_all_neighbors_and_image(structure,rcut,include_index=True)
+    structure.add_site_property(
+        "UC_index", [str(i) for i in range(len(structure.sites))]
+    )
+    neighbours = get_all_neighbors_and_image(structure, rcut, include_index=True)
     nodes = []
-
 
     no_nodes = len(structure.sites)
     if get_halo == True:
[…]
