### 2019-04-30 f49e86a
Added exception for minimal_clusters before torture(), with tests
