+++ crystal_torture/pymatgen_interface.py
@@ -317,10 +317,7 @@ def clusters_from_file(filename, rcut, elements):
 
         return clusters
     else:
-        print(
-            "The element set fed to 'clusters_from_file' is not a subset of the elements in the file"
-        )
-        raise ValueError
+        raise ValueError("The element set fed to 'clusters_from_file' is not a subset of the elements in the file")
 
 
 def clusters_from_structure(structure, rcut, elements):
@@ -367,10 +364,7 @@ def clusters_from_structure(structure, rcut, elements):
 
         return clusters
     else:
-        print(
-            "The element set fed to 'clusters_from_file' is not a subset of the elements in the file"
-        )
-        raise ValueError
+        raise ValueError("The element set fed to 'clusters_from_file' is not a subset of the elements in the file")
 
 
 def graph_from_structure(structure, rcut, elements):
