### 2019-04-30 026e46b
Reformat changes and tests
