+++ tests/test_pymatgen_dopinng.py
@@ -2,76 +2,87 @@ import unittest
 import copy
 from unittest.mock import Mock
 from crystal_torture import Node, Cluster, tort
-from crystal_torture.pymatgen_doping import count_sites, dope_structure, dope_structure_by_no, sort_structure, index_sites
+from crystal_torture.pymatgen_doping import (
+    count_sites,
+    dope_structure,
+    dope_structure_by_no,
+    sort_structure,
+    index_sites,
+)
 from pymatgen import Structure
 from ddt import ddt, data, unpack
 
+
 @ddt
-class PymatgenDopingTestCase( unittest.TestCase ):
+class PymatgenDopingTestCase(unittest.TestCase):
     """ Test simple doping routines"""
 
     def test_count_sites(self):
 
         structure = Structure.from_file("tests/STRUCTURE_FILES/POSCAR_SPINEL.vasp")
-        structure.add_site_property("label",["A"]*8+["B"]*16+["O"]*32)
+        structure.add_site_property("label", ["A"] * 8 + ["B"] * 16 + ["O"] * 32)
 
-        self.assertEqual(count_sites(structure,species={"Mg"}),8)
-        self.assertEqual(count_sites(structure,species={"Al"}),16)
-        self.assertEqual(count_sites(structure,species={"O"}),32)
-        self.assertEqual(count_sites(structure,species={"Li"}),0)
-        self.assertEqual(count_sites(structure,labels={"A"}),8)
-        self.assertEqual(count_sites(structure,labels={"B"}),16)
-        self.assertEqual(count_sites(structure,labels={"O"}),32)
-        self.assertEqual(count_sites(structure,labels={"Li"}),0)
-        self.assertEqual(count_sites(structure,labels={"A","B","O"},species={"Mg","Al","O"}),56)
+        self.assertEqual(count_sites(structure, species={"Mg"}), 8)
+        self.assertEqual(count_sites(structure, species={"Al"}), 16)
+        self.assertEqual(count_sites(structure, species={"O"}), 32)
+        self.assertEqual(count_sites(structure, species={"Li"}), 0)
+        self.assertEqual(count_sites(structure, labels={"A"}), 8)
+        self.assertEqual(count_sites(structure, labels={"B"}), 16)
+        self.assertEqual(count_sites(structure, labels={"O"}), 32)
+        self.assertEqual(count_sites(structure, labels={"Li"}), 0)
+        self.assertEqual(
+            count_sites(structure, labels={"A", "B", "O"}, species={"Mg", "Al", "O"}),
+            56,
+        )
 
-    @data(3,4,5,6)
-    def test_doping(self,value):
+    @data(3, 4, 5, 6)
+    def test_doping(self, value):
 
         structure = Structure.from_file("tests/STRUCTURE_FILES/POSCAR_SPINEL.vasp")
-        structure.add_site_property("label",["A"]*8+["B"]*16+["O"]*32)
+        structure.add_site_property("label", ["A"] * 8 + ["B"] * 16 + ["O"] * 32)
 
         for conc in [0.5]:
             structure_temp = copy.deepcopy(structure)
-            structure_temp.make_supercell([value,value,value])   
-            dope_structure(structure_temp,conc,"Mg",["Li","Al"])
-            doped_conc = count_sites(structure_temp,species={"Li","Al"},labels={"A"})/count_sites(structure_temp,labels={"A"})
-            self.assertEqual(conc,doped_conc)
+            structure_temp.make_supercell([value, value, value])
+            dope_structure(structure_temp, conc, "Mg", ["Li", "Al"])
+            doped_conc = count_sites(
+                structure_temp, species={"Li", "Al"}, labels={"A"}
[…]

[…]
