### 2019-04-30 644cea3
version bump -> 1.0.7
