### 2019-04-30 5886f0a
Bug fix for deprecated species_and_occu in pymatgen
