### 2019-04-30 a0c0e44
added contributing guide
