### 2019-04-30 #1306 @connorourke
@mkhorton - I've added some brief community guidelines (as well as sorting out the other issues you pointed out. Thanks for the tip on `Black` - found that very useful)
