### 2019-05-09 #1306 @katyhuff
I will begin a detailed review, just in case. It should go quickly as I've already read the paper and looked over the repository.
