### 2019-06-04 #1306 @katyhuff
@connorourke I apologize again for the delay in this process. In my review I have found a couple of issues:

- We require that the paper should "clearly state what problems the software is designed to solve and who the target audience is". I found your statement of need less than clear. While the physics and methodological background is very important for understanding the statement of need, perhaps a very concise, clear statement of need (and audience) could be added at the beginning of the paper to satisfy this requirement.
- All of the references need to have DOIs. This paper has one, among a long list of references. Please see the above whedon response regarding the DOIs that are missing from your references.

Once these two issues are solve, we can move forward with this submission. Thank you for your patience.
