### 2019-06-21 826e343
DOIs added for paper refs
