+++ paper/paper.bib
@@ -22,6 +22,8 @@ url={https://doi.org/10.1038/s41598-018-27851-x}
         volume = {68},
         number = {C},
         pages = {314--319},
+        doi = {https://doi.org/10.1016/j.commatsci.2012.10.028},
+        url = {http://www.sciencedirect.com/science/article/pii/S0927025612006295},
         month = feb
 }
 
@@ -32,6 +34,8 @@ url={https://doi.org/10.1038/s41598-018-27851-x}
     journal = {J. Phys. Chem. Solids},
     year = {1957},
     volume = {3},
+    doi = {https://doi.org/10.1016/0022-3697(57)90059-8},
+    url = {http://www.sciencedirect.com/science/article/pii/0022369757900598},
     pages = {121--127}
 }
 
@@ -40,6 +44,8 @@ url={https://doi.org/10.1038/s41598-018-27851-x}
     title = {Computer Simulation Studies of Transport in Solids},
     journal = {Ann. Rev. Mater. Sci.},
     year = {1986},
+    doi = {10.1146/annurev.ms.16.080186.002505},
+     URL = {https://doi.org/10.1146/annurev.ms.16.080186.002505},
     volume = {16},
     pages = {517--548}
 }
@@ -49,6 +55,8 @@ url={https://doi.org/10.1038/s41598-018-27851-x}
     title = {Static Lattice Simulation of Structure and Transport in Superionic Conductors},
     journal = {Sol. Stat. Ionics},
     year = {1983},
+    doi = {https://doi.org/10.1016/0167-2738(83)90069-3},
+    url = {http://www.sciencedirect.com/science/article/pii/0167273883900693},
     volume = {8},
     pages = {89--107}
 }
@@ -59,6 +67,8 @@ url={https://doi.org/10.1038/s41598-018-27851-x}
     journal = {Phys. Rev. Lett.},
     year = {2014},
     volume = {112},
+    doi = {10.1103/PhysRevLett.112.145901},
+    url = {https://link.aps.org/doi/10.1103/PhysRevLett.112.145901},
     pages = {145901}
 }
 
@@ -67,6 +77,8 @@ url={https://doi.org/10.1038/s41598-018-27851-x}
     title = {First-Principles Calculation of Self-Diffusion Coefficients},
     journal = {Phys. Rev. Lett.},
     year = {2008},
+    doi = {10.1103/PhysRevLett.100.215901},
+    url = {https://link.aps.org/doi/10.1103/PhysRevLett.100.215901},
     volume = {100},
     pages = {215901}
 }
@@ -77,6 +89,8 @@ url={https://doi.org/10.1038/s41598-018-27851-x}
     journal = {R. Soc. Open. Sci.},
     year = {2017},
     volume = {4},
+    doi = {10.1098/rsos.170824},
+    URL = {https://royalsocietypublishing.org/doi/abs/10.1098/rsos.170824},
     pages = {170824}
 }
 
@@ -86,6 +100,8 @@ url={https://doi.org/10.1038/s41598-018-27851-x}
     journal = {Acc. Chem. Rev.},
     year = {2013},
     volume = {46},
+    doi = {10.1021/ar200329r},
+    URL = {https://doi.org/10.1021/ar200329r},
     pages = {1216--1225}
 }
 
@@ -94,6 +110,8 @@ url={https://doi.org/10.1038/s41598-018-27851-x}
     title = {Crystallography: A Very Short Introduction},
     publisher = {OUP Oxford},
     year = {2016},
+    doi = {10.1107/S2053273316018829},
+    url = {https://doi.org/10.1107/S2053273316018829},
[…]
