+++ paper/paper.md
@@ -25,7 +25,7 @@ bibliography: paper.bib
 
 `crystal-torture` is a Python, Fortran, and OpenMP module for the analysis of diffusion networks in crystal structures.
 
-The transport of mobile ions through crystalline solids is a fundamental process underlying phenomena such as solid-state reactions, and the behaviour of devices such as batteries and fuel cells. Quantitative descriptions of microscopic ionic transport are often derived by considering ionic trajectories as sequences of discrete &ldquo;hops&rdquo; made by individual ions moving between lattice sites in the host crystal structure [@Vineyard_JPhysChemSolids1957; @Catlow_AnnuRevMaterSci1986; @VanDerVenEtAl_AccChemRes2013; @Morgan_RSocOpenSci2017]. In a perfect crystal structure, the crystal symmetry means the full lattice can be constructed by periodically tiling a unit cell, containing a minimal number of lattice sites [@Glazer_IntroductionBook]. In conventional ionic conductors, ion diffusion on this periodic lattice can be modelled as a stochastic random walk, allowing derivation of simple quantitative relationships between the average microscopic hop rate and macroscopic transport coefficients, such as diffusion coefficient and ionic conductivities [@Catlow_SolStatIonics1983; @Catlow_AnnuRevMaterSci1986; @MorganAndMadden_PhysRevLett2014; @Morgan_RSocOpenSci2017; @MantinaEtAl_PhysRevLett2008]. This model is complicated in highly disordered crystal structures where some proportion of the host lattice sites are occupied by immobile atoms, which block the diffusion of nominally mobile ionic species. The long-ranged diffusion of mobile ions now depends on the proportion and arrangement of blocked sites, and the degree to which the mobile ions can access percolating paths through the crystal structure \[Fig. 1\]  [@GarciaDazaEtAl_ACSApplMaterInter2018; @Rustad_arXiv2016; @DengEtAl_ChemMater2015; @UrbanEtAl_AdvEnergyMater2014; @LeeEtAl_Science2014]. 
[…]
