### 2019-06-21 6a694b7
Paper update more concise need statement
