+++ paper/paper.md
@@ -22,9 +22,12 @@ affiliations:
 date: 12 January 2019
 bibliography: paper.bib
 ---
+# Summary
 
-`crystal-torture` is a Python, Fortran, and OpenMP module for the analysis of diffusion networks in crystal structures.
+`crystal-torture` is a Python, Fortran, and OpenMP module for the analysis of diffusion networks in crystal structures.  Ionic diffusion through crystalline solids depends not only on the dynamics of ions within the crystal, but also the connectivity of the transport network. `crystal-torture` provides a [Python API](https://crystal-torture.readthedocs.io/en/latest/) for interrogating network connectivity and diffusion pathways in partially blocked crystal structures. It can be used as a tool for materials scientists to quickly build up network connectivity statistics to determine the viability of potential ionic conductors, before the use of more computationally expensive approaches modelling the full dynamics.
 
+
+# Scientific Context
 The transport of mobile ions through crystalline solids is a fundamental process underlying phenomena such as solid-state reactions, and the behaviour of devices such as batteries and fuel cells. Quantitative descriptions of microscopic ionic transport are often derived by considering ionic trajectories as sequences of discrete &ldquo;hops&rdquo; made by individual ions moving between lattice sites in the host crystal structure [@Vineyard_JPhysChemSolids1957; @Catlow_AnnuRevMaterSci1986; @VanDerVenEtAl_AccChemRes2013; @Morgan_RSocOpenSci2017]. In a perfect crystal structure, the crystal symmetry means the full lattice can be constructed by periodically tiling a unit cell, containing a minimal number of lattice sites [@Glazer_IntroductionBook]. In conventional ionic conductors, ion diffusion on this periodic lattice can be modelled as a stochastic random walk, allowing derivation of simple quantitative relationships between the average microscopic hop rate and macroscopic transport coefficients, such as diffusion coefficient and ionic conductivities [@Catlow_SolStatIonics1983; @Catlow_AnnuRevMaterSci1986; @MorganAndMadden_PhysRevLett2014; @Morgan_RSocOpenSci2017; @MantinaEtAl_PhysRevLett2008]. This model is complicated in highly disordered crystal structures where some proportion of the host lattice sites are occupied by immobile atoms, which block the diffusion of nominally mobile ionic species. The long-ranged diffusion of mobile ions now depends on the proportion and arrangement of blocked sites, and the degree to which the mobile ions can access percolating paths through the crystal structure \[Fig. 1\]  [@GarciaDazaEtAl_ACSApplMaterInter2018; @DengEtAl_ChemMater2015; @UrbanEtAl_AdvEnergyMater2014; @LeeEtAl_Science2014]. 
 If the proportion of blocked sites, $p_\mathrm{b}$, exceeds $1-p$, where $p$ is the site-percolation threshold for that crystal lattice, then no continuous paths exist and the diffusion coefficient and ionic conductivity for the mobile ions are zero. 
 If the proportion of blocked sites falls below this threshold ($0<p_b<1-p$) at least one percolating path exists and the mobile-ion transport coefficients are non-zero, but are decreased relative to the corresponding values for the ideal fully open lattice ($p_b=0$).
@@ -42,7 +45,7 @@ $$\tau^\mathrm{micro}_i= \frac{\min n_{i\to i^\prime}}{\min n^\mathrm{ideal}_{i\
 
 The microscopic tortuosity is a microscopic analogue of the &ldquo;geometric tortuosity&rdquo; [@Clennell_GeogSocLon1997].
 
[…]
