### 2019-06-21 cd0fc87
Paper update
