### 2019-06-21 #1306 @connorourke
@katyhuff @danielskatz - the latest commit should address the paper changes requested in https://github.com/openjournals/joss-reviews/issues/1306#issuecomment-498647832
