### 2019-06-21 c7287e0
Statement of need added to README
