### 2019-06-24 #1306 @katyhuff
@connorourke : The metadata in the DOI must match the JOSS metadata. Currently, the title of your DOI is "connorourke/crystal_torture: JOSS reviewed release" and the JOSS submission title is "crystal-torture: A crystal tortuosity module". You should be able to edit the zenodo metadata.. Could you please be sure that the DOI metadata (title, authors, etc.) matches this JOSS submission and let me know when that's complete?
