### 2019-06-24 #1306 @katyhuff
I see there's a new version, and the subtitle has changed in that version, but the title seems to have remained the same. This is clear from the heading on the landing page as well as the title in the bibtex export.

![image](https://user-images.githubusercontent.com/393899/60026675-9d814680-9661-11e9-9eb2-017f070d872d.png)

```bib
@misc{conn_o_rourke_2019_3254426,
  author       = {Conn O'Rourke and
                  Benjamin Morgan},
  title        = {connorourke/crystal_torture: JOSS review release},
  month        = jun,
  year         = 2019,
  doi          = {10.5281/zenodo.3254426},
  url          = {https://doi.org/10.5281/zenodo.3254426}
}
```
