### 2019-06-24 #1306 @connorourke
Hi @katyhuff - metadata should now be correct on version:
https://doi.org/10.5281/zenodo.3254426

Let me know if there any more issues.
