### 2019-06-24 #1306 @katyhuff
@connorourke : Is  v1.0.08 the desired version?

- v.1.0.08  is what zenodo says
- v.1.0.08 is also the version on github https://github.com/connorourke/crystal_torture/releases
- but the submission was v.1.0.3
