### 2019-06-24 #1306 @connorourke
Hi @katyhuff, yes v1.0.08 is the correct version. There have been a few changes from the review, so the version has bumped.

Thanks
