### 2019-06-24 #1306 @katyhuff
@openjournals/joss-eics this is ready to accept!
