### 2019-06-24 #1306 @katyhuff
@openjournals/jose-eics This is ready to accept, but I'm not sure what this crossref xml error is about ^^ .
