### 2019-06-24 #1306 @kyleniemeyer
Hmm I can't see any issues with the paper that are obvious to me—it compiled ok a few days ago. @arfon ?
