### 2019-06-24 #1306 @kyleniemeyer
Ah, so it appears the issue is just with CrossRef.
