### 2019-06-24 #1306 @arfon
Just a heads up that I'm working on this. It looks like Crossref is having some issues right now and I'm processing the acceptance manually.
