### 2019-06-24 #1306 @arfon
@kyleniemeyer - this looks to be fixed now:

@mkhorton, @katyhuff - many thanks for your reviews and to @katyhuff for editing this one too ✨
